+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+* On Windows, Shiny Express app files are now read in as UTF-8. (#1203)
+
 ### Other changes
 
 ## [0.8.1] - 2024-03-06